o8.1-Incremental:
At this point we would like to build Weekly
the need of daily is not longer, so we can remove some variables
and make life easier. So far incremental could be done daily
except the day when Full is ran.
while [ day != full ] do ; incremental;  done
or we can do odd or even day based on when full ran.
$ [ $((day%2)) -eq 0 ]

Change-Id: I20a9946fa4a6b2f0ffa125766b9fd39694663b53